pyproject.toml
[project]
name = "home-stable-diffusion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "basicsr>=1.4.2",
    "diffusers>=0.36.0",
    "ftfy>=6.3.1",
    "numpy>=2.4.0",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "safetensors>=0.7.0",
    "scipy>=1.16.3",
    "spandrel>=0.4.1",
    "torch>=2.9.1",
    "torchaudio>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.3",
]
